+++ bosesoundtouchapi/bstconst.py
@@ -4,7 +4,7 @@
 # constants are placed in this file if they are used across multiple files.
 # the only exception to this is for the VERSION constant, which is placed here for convenience.
 
-VERSION:str = "1.0.57"
+VERSION:str = "1.0.58"
 """ 
 Current version of the Bose SoundTouch API Python3 Library. 
 """